# swagger-spec.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/v1/organizations/{id}/courses/{courseId}": {
      "patch": {
        "operationId": "OrganizationsController_updateOrganizationCourse_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "courseId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateOrganizationCourseDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Course updated."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update organization course",
        "tags": [
          "organizations"
        ]
      },
      "delete": {
        "operationId": "OrganizationsController_deleteOrganizationCourse_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "courseId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Course deleted."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Delete organization course",
        "tags": [
          "organizations"
        ]
      }
    },
...
    "/v1/courses": {
      "post": {
        "operationId": "CoursesController_create_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCourseDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Course created."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a new course",
        "tags": [
          "courses"
        ]
      },
      "get": {
        "operationId": "CoursesController_findAll_v1",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "category",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "level",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "beginner",
                "intermediate",
                "advanced"
              ]
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Courses retrieved."
          }
        },
        "summary": "Get all courses",
        "tags": [
          "courses"
        ]
      }
    },
    "/v1/courses/user/purchased": {
      "get": {
        "operationId": "CoursesController_getUserPurchasedCourses_v1",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Purchased courses retrieved."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get user purchased courses",
        "tags": [
          "courses"
        ]
      }
    },
    "/v1/courses/instructor/{instructorId}": {
      "get": {
        "operationId": "CoursesController_getByInstructor_v1",
        "parameters": [
          {
            "name": "instructorId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Instructor courses retrieved."
          }
        },
        "summary": "Get courses by instructor",
        "tags": [
          "courses"
        ]
      }
    },
    "/v1/courses/{id}": {
      "get": {
        "operationId": "CoursesController_findOne_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Course retrieved."
          }
        },
        "summary": "Get course by ID",
        "tags": [
          "courses"
        ]
      },
      "patch": {
        "operationId": "CoursesController_update_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCourseDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Course updated."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update course",
        "tags": [
          "courses"
        ]
      },
      "delete": {
        "operationId": "CoursesController_remove_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Course deleted."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Delete course",
        "tags": [
          "courses"
        ]
      }
    },
    "/v1/courses/{id}/enroll": {
      "post": {
        "operationId": "CoursesController_enroll_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Enrolled successfully."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Enroll in course",
        "tags": [
          "courses"
        ]
      }
    },
    "/v1/courses/{id}/editors": {
      "post": {
        "operationId": "CoursesController_addEditor_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddEditorDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Editor added."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Add editor to course",
        "tags": [
          "courses"
        ]
      }
    },
    "/v1/courses/{id}/editors/{editorId}": {
      "delete": {
        "operationId": "CoursesController_removeEditor_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "editorId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Editor removed."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Remove editor from course",
        "tags": [
          "courses"
        ]
      }
    },
...
    "/v1/discounts/{id}": {
      "get": {
        "operationId": "DiscountController_findOne_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Discount"
        ]
      },
      "patch": {
        "operationId": "DiscountController_update_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Discount"
        ]
      },
      "delete": {
        "operationId": "DiscountController_remove_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Discount"
        ]
      }
    },
    "/v1/discounts/code/{code}": {
      "get": {
        "operationId": "DiscountController_findByCode_v1",
        "parameters": [
          {
            "name": "code",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Discount"
        ]
      }
    },
...
    "/v1/invoices/{id}": {
      "get": {
        "operationId": "InvoiceController_findOne_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Invoice"
        ]
      },
      "patch": {
        "operationId": "InvoiceController_update_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Invoice"
        ]
      },
      "delete": {
        "operationId": "InvoiceController_remove_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Invoice"
        ]
      }
    },
    "/v1/invoices/number/{invoiceNumber}": {
      "get": {
        "operationId": "InvoiceController_findByInvoiceNumber_v1",
        "parameters": [
          {
            "name": "invoiceNumber",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Invoice"
        ]
      }
    },
...
    "/v1/levels/{id}": {
      "get": {
        "operationId": "LevelController_findOne_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Level"
        ]
      },
      "patch": {
        "operationId": "LevelController_update_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Level"
        ]
      },
      "delete": {
        "operationId": "LevelController_remove_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Level"
        ]
      }
    },
    "/v1/levels/organization/{organizationId}": {
      "get": {
        "operationId": "LevelController_findByOrganization_v1",
        "parameters": [
          {
            "name": "organizationId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Level"
        ]
      }
    },
...
    "/v1/materials/course/{courseId}": {
      "get": {
        "operationId": "MaterialController_findByCourse_v1",
        "parameters": [
          {
            "name": "courseId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Material"
        ]
      }
    },
...
  },
  "info": {
    "title": "Shadboard API",
    "description": "The Shadboard API description",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "UpdateOrganizationCourseDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "thumbnailUrl": {
            "type": "string"
          },
          "level": {
            "type": "string"
          },
          "category": {
            "type": "string"
          },
          "price": {
            "type": "number"
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "isPublished": {
            "type": "boolean"
          },
          "isOrgPrivate": {
            "type": "boolean"
          }
        }
      },
...
      "CreateCourseDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "thumbnailUrl": {
            "type": "string"
          },
          "level": {
            "type": "string",
            "enum": [
              "beginner",
              "intermediate",
              "advanced"
            ]
          },
          "category": {
            "type": "string"
          },
          "price": {
            "type": "number"
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "isPublished": {
            "type": "boolean"
          },
          "organizationId": {
            "type": "string"
          },
          "isOrgPrivate": {
            "type": "boolean"
          }
        },
        "required": [
          "title",
          "description",
          "level",
          "category",
          "price"
        ]
      },
      "UpdateCourseDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "thumbnailUrl": {
            "type": "string"
          },
          "level": {
            "type": "string",
            "enum": [
              "beginner",
              "intermediate",
              "advanced"
            ]
          },
          "category": {
            "type": "string"
          },
          "price": {
            "type": "number"
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "isPublished": {
            "type": "boolean"
          },
          "isOrgPrivate": {
            "type": "boolean"
          }
        }
      },
      "AddEditorDto": {
        "type": "object",
        "properties": {
          "editorId": {
            "type": "string"
          }
        },
        "required": [
          "editorId"
        ]
      },
...
    }
  }
}